+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Moves the edit signal button in the popup menu button in the app bar.
+- Initializes the calendar with the current day around the middle of the screen.
 
 ## 0.16.0